[CODE] Basic working harness
Made a very basic harness. Got it to work with the abstract data generator.

I'm a little concerned this new structure is too cumbersome. It might suck to have to deal with getting disorganized and messy, but worth it to get rid of the structure for flexibility. Maybe keep some structure, but make it a lot less?